Moved middleware marshaling methods to a trait
- Allows re-use for alternate routing/dispatching implementations.
- Reduces immediate LOC in Application class.
